# Conflicts:
#	Project Peter's Band
UE4/Content/FinalGameLevel/SectionFour/E22_Lobby.umap

+++ Project Peter's Band UE4/Source/FirstPersonInput/Lights/MemorySegmentLights.h
@@ -66,5 +66,5 @@ public:
 	FName MapToTeleportTo;
 
 	UPROPERTY(EditAnywhere)
-	uint8 ExtraTimeBuffer;
+	float ExtraTimeBuffer;
 };
